Fix for Travis build, now requirements.txt on master is for python3

+++ bin/travis-build.bash
@@ -20,7 +20,12 @@ fi
 sed -i '/psycopg2/c\psycopg2' requirements.txt
 
 python setup.py develop
-pip install -r requirements.txt
+if [ -f requirements-py2.txt ]
+then
+    pip install -r requirements-py2.txt
+else
+    pip install -r requirements.txt
+fi
 pip install -r dev-requirements.txt
 cd -
